Rename Backend trait to DistributeFile
